follow along
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6810 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/RecordReplay.cxx
@@ -534,7 +534,7 @@ int Record::getSize ()
 
 int Record::getRate ()
 {
-  return (RecordUpdateRate / (RRtime)1000000);
+  return (int)(RecordUpdateRate / (RRtime)1000000);
 }
